Refactor CodesplainAppbar to logout users with invalid tokens and refactor promise logic
